1. Event Detail Screen Appearance improved. 2. Speakers Screen ready. 3. Speaker ProfileScreen in progress (need to improve DB for relation with Events)
[completed: 14]{t:90}
[completed: 12]{t:180}
[touch: 17]{t:180}

+++ DrupalCon/DrupalCon/Classes/ConferenceViewControllers/Program/DCProgramItemsViewController.m
@@ -120,7 +120,7 @@
     return cell;
 }
 
--(BOOL) headerNeededInSection: (int) section
+-(BOOL) headerNeededInSection: (NSInteger) section
 {
     /*lets check if this date range contains some events that need a time period header, DCSpeechCelll and DCSPeechofTheDayCell, if its only coffe breaks or lunch - we dont display a header*/
     BOOL headerNeeded = NO;
